+++ src/Aguacongas.Firebase/FirebaseClient.cs
@@ -7,29 +7,35 @@ using System.Threading.Tasks;
 
 namespace Aguacongas.Firebase
 {
-    public class FirebaseClient : IFirebaseClient, IDisposable
+    /// <summary>
+    /// Firebase client
+    /// </summary>
+    public class FirebaseClient : IFirebaseClient
     {
         private readonly HttpClient _httpClient;
-        private readonly string _url;
-        private readonly IFirebaseTokenManager _tokenManager;
         private readonly JsonSerializerSettings _jsonSerializerSettings;
 
-        public FirebaseClient(HttpClient httpClient, IFirebaseTokenManager tokenManager, IOptions<FirebaseOptions> options):
-            this(httpClient, tokenManager, options?.Value?.DatabaseUrl, options?.Value?.JsonSerializerSettings)
-        { }
-
-        public FirebaseClient(HttpClient httpClient, IFirebaseTokenManager tokenManager, string url, JsonSerializerSettings jsonSerializerSettings)
+        /// <summary>
+        /// Create a new instance of <see cref="FirebaseClient"/>
+        /// </summary>
+        /// <param name="httpClient">The <see cref="HttpClient"/></param>
+        /// <param name="options">A <see cref="FirebaseOptions"/></param>
+        public FirebaseClient(HttpClient httpClient, FirebaseOptions options)
         {
             _httpClient = httpClient ?? throw new ArgumentNullException(nameof(httpClient));
-            _tokenManager = tokenManager ?? throw new ArgumentNullException(nameof(tokenManager));
-            _url = url ?? throw new ArgumentNullException(nameof(url));
-
-            if (!_url.EndsWith("/"))
-            {
-                _url = _url + "/";
-            }
+            _httpClient.BaseAddress = new Uri(options.DatabaseUrl);
+            _jsonSerializerSettings = options.JsonSerializerSettings;
         }
 
+        /// <summary>
+        /// Sends a POST request to create data
+        /// </summary>
+        /// <typeparam name="T">Type of data</typeparam>
+        /// <param name="url">Data uri</param>
+        /// <param name="data">Data</param>
+        /// <param name="cancellationToken">A <see cref="CancellationToken"/></param>
+        /// <param name="requestEtag">True to request ETag</param>
+        /// <returns>The data id</returns>
         public async Task<FirebaseResponse<string>> PostAsync<T>(string url, T data, CancellationToken cancellationToken = default(CancellationToken), bool requestEtag = false)
         {
             if(data == null)
@@ -45,6 +51,16 @@ namespace Aguacongas.Firebase
             };
         }
 
+        /// <summary>
+        /// Send a PUT request to update data
+        /// </summary>
+        /// <typeparam name="T">Type of data</typeparam>
+        /// <param name="url">Data uri</param>
+        /// <param name="data">Data</param>
+        /// <param name="cancellationToken">A <see cref="CancellationToken"/></param>
+        /// <param name="requestEtag">True to request ETag</param>
+        /// <param name="etag">ETag value</param>
+        /// <returns>A <see cref="FirebaseResponse{T}"/></returns>
         public Task<FirebaseResponse<T>> PutAsync<T>(string url, T data, CancellationToken cancellationToken = default(CancellationToken), bool requestEtag = false, string etag = null)
         {
             if (data == null)
@@ -55,6 +71,14 @@ namespace Aguacongas.Firebase
             return SendFirebaseRequest<T>(url, "PUT", _httpClient.CreateJsonContent(data, _jsonSerializerSettings), cancellationToken, requestEtag, etag);
         }
 
+        /// <summary>
+        /// Send a PATCH request to path data
+        /// </summary>
+        /// <typeparam name="T">Type of data</typeparam>
+        /// <param name="url">Data uri</param>
+        /// <param name="data">Data</param>
+        /// <param name="cancellationToken">A <see cref="CancellationToken"/></param>
+        /// <returns>A <see cref="FirebaseResponse{T}"/></returns>
         public Task<FirebaseResponse<T>> PatchAsync<T>(string url, T data, CancellationToken cancellationToken = default(CancellationToken))
         {
             if (data == null)
@@ -65,9 +89,17 @@ namespace Aguacongas.Firebase
             return SendFirebaseRequest<T>(url, "PATCH", _httpClient.CreateJsonContent(data, _jsonSerializerSettings), cancellationToken, false, null);
         }
 
+        /// <summary>
+        /// Send a DELETE request to delete data
+        /// </summary>
+        /// <param name="url">Data uri</param>
+        /// <param name="cancellationToken">A <see cref="CancellationToken"/></param>
+        /// <param name="requestEtag">True to request ETag</param>
+        /// <param name="etag">Etag value</param>
+        /// <returns></returns>
         public async Task DeleteAsync(string url, CancellationToken cancellationToken = default(CancellationToken), bool requestEtag = false, string etag = null)
         {
-            var message = new HttpRequestMessage(new HttpMethod("DELETE"), await GetFirebaseUrl(url, cancellationToken));
+            var message = new HttpRequestMessage(new HttpMethod("DELETE"), GetFirebaseUrl(url, cancellationToken));
             message.Headers.SetIfMath(etag);
             message.Headers.SetRequestEtag(requestEtag);
 
@@ -78,6 +110,15 @@ namespace Aguacongas.Firebase
             }                
         }
 
+        /// <summary>
+        /// Send a GET request to request data
+        /// </summary>
+        /// <typeparam name="T">The type of data</typeparam>
+        /// <param name="url">Data uri</param>
+        /// <param name="cancellationToken">A <see cref="CancellationToken"/></param>
+        /// <param name="requestEtag">True to request ETag</param>
+        /// <param name="queryString">A query string</param>
+        /// <returns>A <see cref="FirebaseResponse{T}"/></returns>
         public Task<FirebaseResponse<T>> GetAsync<T>(string url, CancellationToken cancellationToken = default(CancellationToken), bool requestEtag = false, string queryString = null)
         {
             return SendFirebaseRequest<T>(url, "GET", null, cancellationToken, requestEtag, null, queryString);
@@ -85,7 +126,7 @@ namespace Aguacongas.Firebase
 
         private async Task<FirebaseResponse<T>> SendFirebaseRequest<T>(string url, string method, HttpContent content, CancellationToken cancellationToken, bool requestEtag, string eTag, string queryString = null)
         {
-            var message = new HttpRequestMessage(new HttpMethod(method), await GetFirebaseUrl(url, cancellationToken, queryString))
+            var message = new HttpRequestMessage(new HttpMethod(method), GetFirebaseUrl(url, cancellationToken, queryString))
             {
                 Content = content
             };
@@ -95,6 +136,7 @@ namespace Aguacongas.Firebase
             var response = await _httpClient.SendAsync(message, cancellationToken);
             return await GetFirebaseResponse<T>(response);
         }
+
         private async Task<FirebaseResponse<T>> GetFirebaseResponse<T>(HttpResponseMessage response)
         {
             using (response)
@@ -103,7 +145,7 @@ namespace Aguacongas.Firebase
             }
         }
 
-        private async Task<string> GetFirebaseUrl(string url, CancellationToken cancellationToken, string queryString = null)
+        private Uri GetFirebaseUrl(string url, CancellationToken cancellationToken, string queryString = null)
         {
             if (string.IsNullOrEmpty(url))
             {
@@ -116,46 +158,19 @@ namespace Aguacongas.Firebase
                 sanetizedUrl = url + ".json";
             }
 
-            while (sanetizedUrl.StartsWith("/"))
+            var baseAddress = _httpClient.BaseAddress;
+            var builder = new UriBuilder(baseAddress.Scheme, baseAddress.Host, baseAddress.Port, sanetizedUrl)
             {
-                sanetizedUrl = sanetizedUrl.Substring(1);
-            }
+                Query = queryString
+            };
 
-            var result = $"{_url}{sanetizedUrl}?{_tokenManager.AuthParamName}={await _tokenManager.GetTokenAsync(cancellationToken)}";
-            if(!string.IsNullOrEmpty(queryString))
-            {
-                result += "&" + queryString; 
-            }
-            return result;
+            
+            return builder.Uri;
         }
 
         class PostResponse
         {
             public string Name { get; set; }
         }
-
-        #region IDisposable Support
-        private bool disposedValue = false; // To detect redundant calls
-
-        protected virtual void Dispose(bool disposing)
-        {
-            if (!disposedValue)
-            {
-                if (disposing)
-                {
-                    _httpClient.Dispose();
-                    _tokenManager.Dispose();
-                }
-
-                disposedValue = true;
-            }
-        }
-
-        // This code added to correctly implement the disposable pattern.
-        public void Dispose()
-        {
-            Dispose(true);
-        }
-        #endregion
     }
 }