+++ src/Aguacongas.Firebase/FirebaseResponse.cs
@@ -1,13 +1,19 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Aguacongas.Firebase
+﻿namespace Aguacongas.Firebase
 {
+    /// <summary>
+    /// Firebase response
+    /// </summary>
+    /// <typeparam name="T">Type of data</typeparam>
     public class FirebaseResponse<T>
     {
+        /// <summary>
+        /// Gets or sets data
+        /// </summary>
         public T Data { get; set; }
 
+        /// <summary>
+        /// Gets or sets Etag
+        /// </summary>
         public string Etag { get; set; }
     }
 }