+++ src/Aguacongas.Identity.Firebase/LoginIndex.cs
@@ -1,12 +1,13 @@
 ﻿using Aguacongas.Firebase;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace Aguacongas.Identity.Firebase
 {
+    /// <summary>
+    /// Login indexes
+    /// </summary>
     public class LoginIndex : FirebaseIndexes
     {
+        /// <inheritdoc />
         public LoginIndex()
         {
             On = new string[] { "ProviderKey", "UserId" };