+++ src/Aguacongas.Identity.Firebase/RoleClaimIndex.cs
@@ -2,8 +2,12 @@
 
 namespace Aguacongas.Identity.Firebase
 {
+    /// <summary>
+    /// Role claim index
+    /// </summary>
     public class RoleClaimIndex : FirebaseIndex
     {
+        /// <inheritdoc />
         public RoleClaimIndex()
         {
             On = "RoleId";