+++ src/Aguacongas.Identity.Firebase/UserOnlyStore.cs
@@ -82,7 +82,7 @@ namespace Aguacongas.Identity.Firebase
                 var userDictionary = response.Data;
                 if (userDictionary == null)
                 {
-                    return new List<TUser>().AsQueryable();
+                    return new List<TUser>(0).AsQueryable();
                 }
 
                 var taskList = new List<Task<TUser>>();
@@ -91,9 +91,12 @@ namespace Aguacongas.Identity.Firebase
                     taskList.Add(FindByIdAsync(key));
                 }
 
-                var results = Task.WhenAll(taskList).GetAwaiter().GetResult();
+                Task.WaitAll(taskList.ToArray());
+                var users = taskList.Where(t => t.Result != null)
+                    .Select(t => t.Result)
+                    .ToList();
 
-                return results.Where(u => u != null).AsQueryable();
+                return users.AsQueryable();
             }
         }