+++ test/Aguacongas.Identity.Firebase.IntegrationTest/FirebaseTestFixture.cs
@@ -1,12 +1,7 @@
 ﻿using Aguacongas.Firebase;
-using Aguacongas.Firebase.TokenManager;
 using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
 using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Net.Http;
-using System.Text;
 
 namespace Aguacongas.Identity.Firebase.IntegrationTest
 {
@@ -21,7 +16,7 @@ namespace Aguacongas.Identity.Firebase.IntegrationTest
             var builder = new ConfigurationBuilder();
             Configuration = builder.AddUserSecrets<UserStoreTest>()
                 .AddEnvironmentVariables()
-                .AddJsonFile(Path.Combine(Directory.GetCurrentDirectory(), @"..\..\..\..\testsettings.json"))
+                .AddJsonFile(Path.Combine(Directory.GetCurrentDirectory(), "../../../../testsettings.json"))
                 .Build();
 
             FirebaseOptions = new FirebaseOptions();