+++ test/Aguacongas.Identity.Firebase.Test/IdentityBuilderExtensionsTest.cs
@@ -16,12 +16,12 @@ namespace Aguacongas.Identity.Firebase.Test
             var builder = new ConfigurationBuilder();
             var configuration = builder.AddUserSecrets<IdentityBuilderExtensionsTest>()
                 .AddEnvironmentVariables()
-                .AddJsonFile(Path.Combine(Directory.GetCurrentDirectory(), @"..\..\..\..\testsettings.json"))
+                .AddJsonFile(Path.Combine(Directory.GetCurrentDirectory(), "../../../../testsettings.json"))
                 .Build();
 
             var services = new ServiceCollection();
             services.AddIdentity<IdentityUser, IdentityRole>()
-                .AddFirebaseStores("http:test", options =>
+                .AddFirebaseStores("http://test", options =>
                 {
                     configuration.GetSection("AuthTokenOptions").Bind(options);
                 });